+++ crates/burn-cubecl-fusion/src/elemwise/builder.rs
@@ -27,8 +27,7 @@ impl<R: Runtime> ElementWiseBuilder<R> {
                 FuseSettings {
                     broadcast: true,
                     output_shape_updates: true,
-                    mix_vectorization: true,
-                    inplace: false,
+                    inplace: true,
                 },
             ),
             device,
@@ -38,12 +37,10 @@ impl<R: Runtime> ElementWiseBuilder<R> {
 
 impl<R: Runtime> OptimizationBuilder<CubeOptimization<R>> for ElementWiseBuilder<R> {
     fn register(&mut self, operation: &burn_ir::OperationIr) {
-        println!("{operation:?}");
         self.builder.register(operation);
     }
 
     fn build(&self) -> CubeOptimization<R> {
-        println!("BUILDING OPE {}", self.len());
         let client = R::client(&self.device);
         let trace = self.builder.build();
         let elementwise =
