+++ crates/burn-cubecl-fusion/src/on_write/settings.rs
@@ -10,11 +10,6 @@ pub struct FuseSettings {
     /// When broadcast is enabled, the output shape can become bigger after a fusion,
     /// therefore an update is needed.
     pub output_shape_updates: bool,
-    /// Enables mix vectorization factor.
-    ///
-    /// Useful when the last dimension is broadcasted for one of the tensors, which would limit the
-    /// vectorization factor to be 1 without this setting enabled.
-    pub mix_vectorization: bool,
     /// Enables the reuse of input buffers.
     pub inplace: bool,
 }
